requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
SQLAlchemy==1.4.44
Werkzeug==2.2.2
zipp==3.11.0
pandas==1.5.1
plotly==5.11.0